# Make SecureSecurityPolicyConfig significantly faster